Improve Unit handling
Move DatalabUnitTransform into `models.units`
Rename `DatalabQuantityExtra` to `DatalabQuantity`. Update imports, regenerate schemas, adjust tests.
